+++ script/signin.js
@@ -1,14 +1,20 @@
-const formEl = document.querySelector("form");
-const emailInputEl = document.querySelector("#email");
-const pwInputEl = document.querySelector("#password");
-const emailErrorMessage = document.querySelector("#email_error");
-const pwErrorMessage = document.querySelector("#pw_error");
-const submitButtonEl = document.querySelector("#formSubmit");
-const pwShowButtonEl = document.querySelector("#eyePw");
-
-const emailRegex =
-    /^(([^<>()\[\].,;:\s@"]+(\.[^<>()\[\].,;:\s@"]+)*)|(".+"))@(([^<>()[\].,;:\s@"]+\.)+[^<>()[\].,;:\s@"]{2,})$/i;
-const pwRegex = /^(?=.*[a-zA-Z])(?=.*[0-9]).{8,25}$/;
+// * as "something" 이런식으로 가져오기도 하는 것 같은데
+// 저는 가져온 변수들이 이 페이지에서 보였으면 좋겠어서 일일이 다 불러오게 되었습니다.
+// 새로운 이름 지정해주고 객체로 뽑아오면 자꾸 파일 왔다갔다하면서 뭐가 있는지 확인하는게 귀찮을 것 같았어요.
+// 어떤 방식이 더 나을까요?
+import {
+    formEl,
+    emailInputEl,
+    pwInputEl,
+    emailErrorMessage,
+    pwErrorMessage,
+    pwShowButtonEl,
+    submitButtonEl,
+    emailRegex,
+    pwRegex,
+} from "./common.js";
+// let은 값이 변경될 수 있으니까 따로 export 해주지 않았습니다.
+// 하지만 해당 변수들도 signin/signup에서 반복되는데 export 해주는게 좋을까요?
 let emailValid = false;
 let pwValid = false;
 let emailInputValue = "";
