+++ script/signup.js
@@ -1,17 +1,17 @@
-const formEl = document.querySelector("form");
-const emailInputEl = document.querySelector("#email");
-const pwInputEl = document.querySelector("#password");
-const pwConfirmInputEl = document.querySelector("#password_confirm");
-const emailErrorMessage = document.querySelector("#email_error");
-const pwErrorMessage = document.querySelector("#pw_error");
-const pwConfirmErrorMessage = document.querySelector("#pwConfirm_error");
-const submitButtonEl = document.querySelector("#formSubmit");
-const pwShowButtonEl = document.querySelector("#eyePw");
-const pwConfirmShowButtonEl = document.querySelector("#eyePwConfirm");
-
-const emailRegex =
-    /^(([^<>()\[\].,;:\s@"]+(\.[^<>()\[\].,;:\s@"]+)*)|(".+"))@(([^<>()[\].,;:\s@"]+\.)+[^<>()[\].,;:\s@"]{2,})$/i;
-const pwRegex = /^(?=.*[a-zA-Z])(?=.*[0-9]).{8,25}$/;
+import {
+    formEl,
+    emailInputEl,
+    pwInputEl,
+    emailErrorMessage,
+    pwErrorMessage,
+    pwShowButtonEl,
+    submitButtonEl,
+    pwConfirmInputEl,
+    emailRegex,
+    pwRegex,
+    pwConfirmErrorMessage,
+    pwConfirmShowButtonEl,
+} from "./common.js";
 let emailValid = false;
 let pwValid = false;
 let pwConfirmValid = false;
