+++ signin.html
@@ -70,6 +70,6 @@
                 </div>
             </div>
         </section>
-        <script type="module" src="/script/signin.js"></script>
+        <script src="/script/signin.js" type="module"></script>
     </body>
 </html>
