+++ signup.html
@@ -86,6 +86,6 @@
                 </div>
             </div>
         </section>
-        <script type="module" src="/script/signup.js"></script>
+        <script src="/script/signup.js" type="module"></script>
     </body>
 </html>
